3.2 CNY Bug and Error in Work Hours
The CNY Bug is caused by an absence of skip day counter in line 202.
Before the modification, when a record finally has no skip, the original
skip day counter is reset but the overnight check is not. Under the
right circumstance where there is an overnight work and no work in the
following day, the date contributed for overnight work would be mistaken
as record for the next day and a duplicated date copy is inserted into
the final results. Under this erroneous entry, this can lead the program
to loop though continuously and the loop protection cannot cover this
situation.
To mitigate the issue, the overnight check is also reset together with
skip day counts.

The work hours error is caused by a mistake in the coding, it should be
discovered in earlier versions.